+++ src/core/converter/other/MsfsPlnToAeroflyFlightConverter.ts
@@ -118,12 +118,17 @@ export class MsfsPlnToAeroflyFlightConverter extends StringToAeroflyFlightConver
         lat: number;
         altitude_ft: number;
     } {
+        if (coordinate === "") {
+            throw new Error(
+                `Missing coordinates in file. Possibly you are trying to import an Microsoft Flight Simulator 2024 EFB file instead of a mission file.`,
+            );
+        }
         // N52° 45' 7.51",W3° 53' 2.16",+002500.00
 
         const parts = coordinate.split(/,\s*/);
         if (parts.length < 2) {
             throw new Error(
-                `Wrong coordinates format "${coordinate}", expexted something like N52° 45' 7.51",W3° 53' 2.16",+002500.00`,
+                `Wrong coordinates format "${coordinate}", expected format like "N52° 45' 7.51",W3° 53' 2.16",+002500.00"`,
             );
         }
         const numbers = parts.map((p): number => {